+++ packages/fx-core/src/core/FxCore.ts
@@ -780,6 +780,42 @@ export class FxCore implements v3.ICore {
         ) {
           await ensureBasicFolderStructure(inputs);
         }
+        // reset provisionSucceeded state for all env
+        if (res.isOk() && (func.method === "addCapability" || func.method === "addResource")) {
+          if (
+            ctx.envInfoV2?.state?.solution?.provisionSucceeded === true ||
+            ctx.envInfoV2?.state?.solution?.provisionSucceeded === "true"
+          ) {
+            ctx.envInfoV2.state.solution.provisionSucceeded = false;
+          }
+          const allEnvRes = await environmentManager.listRemoteEnvConfigs(inputs.projectPath!);
+          if (allEnvRes.isOk()) {
+            for (const env of allEnvRes.value) {
+              const loadEnvRes = await loadEnvInfoV3(
+                inputs as v2.InputsWithProjectPath,
+                ctx.projectSettings!,
+                env,
+                false
+              );
+              if (loadEnvRes.isOk()) {
+                const envInfo = loadEnvRes.value;
+                if (
+                  envInfo.state?.solution?.provisionSucceeded === true ||
+                  envInfo.state?.solution?.provisionSucceeded === "true"
+                ) {
+                  envInfo.state.solution.provisionSucceeded = false;
+                  await environmentManager.writeEnvState(
+                    envInfo.state,
+                    inputs.projectPath!,
+                    ctx.contextV2.cryptoProvider,
+                    env,
+                    true
+                  );
+                }
+              }
+            }
+          }
+        }
         return res;
       } else return err(FunctionRouterError(func));
     }
@@ -1356,7 +1392,7 @@ export class FxCore implements v3.ICore {
     // create folder structure
     await fs.ensureDir(path.join(projectPath, `.${ConfigFolderName}`));
     await fs.ensureDir(path.join(projectPath, "templates", `${AppPackageFolderName}`));
-    const basicFolderRes = await ensureBasicFolderStructure(inputs);
+    const basicFolderRes = await ensureBasicFolderStructure(inputs, false);
     if (basicFolderRes.isErr()) {
       return err(basicFolderRes.error);
     }
@@ -1450,42 +1486,54 @@ export class FxCore implements v3.ICore {
   _getQuestionsForUserTaskV3 = getQuestionsForUserTaskV3;
 }
 
-export async function ensureBasicFolderStructure(inputs: Inputs): Promise<Result<null, FxError>> {
+export async function ensureBasicFolderStructure(
+  inputs: Inputs,
+  createPackageJson = true
+): Promise<Result<null, FxError>> {
   if (!inputs.projectPath) {
     return err(new ObjectIsUndefinedError("projectPath"));
   }
   try {
-    // {
-    //   const appName = inputs[CoreQuestionNames.AppName] as string;
-    //   if (inputs.platform !== Platform.VS) {
-    //     const packageJsonFilePath = path.join(inputs.projectPath, `package.json`);
-    //     const exists = await fs.pathExists(packageJsonFilePath);
-    //     if (!exists) {
-    //       await fs.writeFile(
-    //         packageJsonFilePath,
-    //         JSON.stringify(
-    //           {
-    //             name: appName,
-    //             version: "0.0.1",
-    //             description: "",
-    //             author: "",
-    //             scripts: {
-    //               test: 'echo "Error: no test specified" && exit 1',
-    //             },
-    //             devDependencies: {
-    //               "@microsoft/teamsfx-cli": "0.*",
-    //             },
-    //             license: "MIT",
-    //           },
-    //           null,
-    //           4
-    //         )
-    //       );
-    //     }
-    //   }
-    // }
+    if (createPackageJson) {
+      const appName = inputs[CoreQuestionNames.AppName] as string;
+      if (inputs.platform !== Platform.VS) {
+        const packageJsonFilePath = path.join(inputs.projectPath, `package.json`);
+        const exists = await fs.pathExists(packageJsonFilePath);
+        if (!exists) {
+          await fs.writeFile(
+            packageJsonFilePath,
+            JSON.stringify(
+              {
+                name: appName,
+                version: "0.0.1",
+                description: "",
+                author: "",
+                scripts: {
+                  test: 'echo "Error: no test specified" && exit 1',
+                },
+                devDependencies: {
+                  "@microsoft/teamsfx-cli": "0.*",
+                },
+                license: "MIT",
+              },
+              null,
+              4
+            )
+          );
+        }
+      }
+    }
     {
       const gitIgnoreFilePath = path.join(inputs.projectPath, `.gitignore`);
+      let lines: string[] = [];
+      const exists = await fs.pathExists(gitIgnoreFilePath);
+      if (exists) {
+        const content = await fs.readFile(gitIgnoreFilePath, { encoding: "utf8" });
+        lines = content.split("\n");
+        for (let i = 0; i < lines.length; ++i) {
+          lines[i] = lines[i].trim();
+        }
+      }
       const gitIgnoreContent = [
         "\n# TeamsFx files",
         "node_modules",
@@ -1496,17 +1544,19 @@ export async function ensureBasicFolderStructure(inputs: Inputs): Promise<Result
         "subscriptionInfo.json",
         BuildFolderName,
       ];
-
       if (isConfigUnifyEnabled()) {
         gitIgnoreContent.push(`.${ConfigFolderName}/${InputConfigsFolderName}/config.local.json`);
         gitIgnoreContent.push(`.${ConfigFolderName}/${StatesFolderName}/state.local.json`);
       }
-
       if (inputs.platform === Platform.VS) {
         gitIgnoreContent.push("appsettings.Development.json");
       }
-
-      await fs.appendFile(gitIgnoreFilePath, gitIgnoreContent.join("\n"), { encoding: "utf8" });
+      gitIgnoreContent.forEach((line) => {
+        if (!lines.includes(line.trim())) {
+          lines.push(line.trim());
+        }
+      });
+      await fs.writeFile(gitIgnoreFilePath, lines.join("\n"), { encoding: "utf8" });
     }
   } catch (e) {
     return err(WriteFileError(e));