+++ packages/fx-core/src/plugins/resource/spfx/utils/telemetryEvents.ts
@@ -4,8 +4,11 @@
 export enum TelemetryEvents {
   EnsureYoStart = "ensure-yo-start",
   EnsureYo = "ensure-yo",
+  EnsureSharepointGeneratorStart = "ensure-sharepoint-generator-start",
+  EnsureSharepointGenerator = "ensure-sharepoint-generator",
 }
 
 export enum TelemetryProperty {
   EnsureYoReason = "ensure-yo-reason",
+  EnsureSharepointGeneratorReason = "ensure-sharepoint-generator-reason",
 }