+++ packages/vscode-extension/src/handlers.ts
@@ -1143,7 +1143,17 @@ export async function validateLocalPrerequisitesHandler(): Promise<string | unde
 export async function installAppInTeams(): Promise<string | undefined> {
   let shouldContinue = false;
   try {
-    shouldContinue = await showInstallAppInTeamsMessage(false);
+    const debugConfig = isConfigUnifyEnabled()
+      ? await commonUtils.getDebugConfig(false, environmentManager.getLocalEnvName())
+      : await commonUtils.getDebugConfig(true);
+    if (debugConfig?.appId === undefined) {
+      throw returnUserError(
+        new Error("Debug config not found"),
+        ExtensionSource,
+        ExtensionErrors.GetTeamsAppInstallationFailed
+      );
+    }
+    shouldContinue = await showInstallAppInTeamsMessage(false, debugConfig.appId);
   } catch (error: any) {
     showError(error);
   }
@@ -1354,9 +1364,9 @@ function getTriggerFromProperty(args?: any[]) {
 }
 
 async function autoOpenProjectHandler(): Promise<void> {
-  const isOpenWalkThrough = globalStateGet(GlobalKey.OpenWalkThrough, false);
-  const isOpenReadMe = globalStateGet(GlobalKey.OpenReadMe, false);
-  const isOpenSampleReadMe = globalStateGet(GlobalKey.OpenSampleReadMe, false);
+  const isOpenWalkThrough = await globalStateGet(GlobalKey.OpenWalkThrough, false);
+  const isOpenReadMe = await globalStateGet(GlobalKey.OpenReadMe, false);
+  const isOpenSampleReadMe = await globalStateGet(GlobalKey.OpenSampleReadMe, false);
   if (isOpenWalkThrough) {
     showLocalDebugMessage();
     await openWelcomeHandler([TelemetryTiggerFrom.Auto]);
@@ -1459,7 +1469,7 @@ async function postUpgrade(): Promise<void> {
 
 async function popupAfterUpgrade(): Promise<void> {
   const aadClientSecretFlag = "NeedToSetAADClientSecretEnv";
-  const aadClientSecret = globalStateGet(aadClientSecretFlag, "");
+  const aadClientSecret = await globalStateGet(aadClientSecretFlag, "");
   if (
     aadClientSecret !== "" &&
     workspace.workspaceFolders &&
@@ -1493,7 +1503,7 @@ async function popupAfterUpgrade(): Promise<void> {
 async function openUpgradeChangeLogsHandler() {
   const openUpgradeChangelogsFlag = "openUpgradeChangelogs";
   if (
-    globalStateGet(openUpgradeChangelogsFlag, false) &&
+    (await globalStateGet(openUpgradeChangelogsFlag, false)) &&
     workspace.workspaceFolders &&
     workspace.workspaceFolders.length > 0
   ) {
@@ -1538,7 +1548,7 @@ async function openSampleReadmeHandler(args?: any) {
 }
 
 async function showLocalDebugMessage() {
-  const isShowLocalDebugMessage = globalStateGet(GlobalKey.ShowLocalDebugMessage, false);
+  const isShowLocalDebugMessage = await globalStateGet(GlobalKey.ShowLocalDebugMessage, false);
 
   if (!isShowLocalDebugMessage) {
     return;
@@ -2699,4 +2709,15 @@ export async function openDeploymentTreeview(args?: any[]) {
   } else {
     vscode.commands.executeCommand("workbench.view.extension.teamsfx");
   }
+}
+
+export async function addSsoHanlder(): Promise<Result<null, FxError>> {
+  ExtTelemetry.sendTelemetryEvent(TelemetryEvent.AddSsoStart);
+  const func: Func = {
+    namespace: "fx-solution-azure",
+    method: "addSso",
+  };
+
+  const result = await runUserTask(func, TelemetryEvent.AddSso, true);
+  return result;
 }